requirements.txt
asgiref==3.2.7
Django==3.0.6
django-tinymce4-lite==1.8.0
jsmin==2.2.2
pyperclip==1.8.1
pytz==2020.1
sqlparse==0.3.1
